backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pymupdf
faiss-cpu
numpy
requests